core-db/providers/web: #435 small provable trims
Five small mechanical shrinks, zero behavior change:

- session.py: ensure_active_session now delegates to get_current_session
  for stale-pointer validation instead of hand-rolling the same
  read-pointer -> open-db -> check-row -> clear-stale sequence.
- db/upgrades.py: drop stale pre-v0.9.0 scaffolding COMMENTS (the
  held-at-8/DORMANT/xfail paragraph and the authoring-rule docstring
  block, now condensed to a GH-0352 pointer). The v4-v6 (and v8->v9)
  upgrade STEPS themselves stay untouched.
- benchmark/clients.py: folded its single make_openai_client function
  into stores.py (next to BenchmarkRoot) and deleted the module; the
  two lazy-import callers (summarize, judging) updated.
- web queries.js: listTags() now derives from listAllTags() with a
  >0-doc filter instead of a second near-identical SQL read.
- web queries.js: getCitations() stops selecting/returning the text and
  source_kind fields no consumer reads.

Part of #447.
